+++ src/shadow/texture-shader.ts
@@ -1,5 +1,4 @@
-import * as PIXI from "pixi.js"
-
+import { Renderer, Buffer } from "pixi.js"
 import { MeshGeometry3D } from "../mesh/geometry/mesh-geometry"
 import { Mesh3D } from "../mesh/mesh"
 import { ShadowCastingLight } from "./shadow-casting-light"
@@ -11,15 +10,11 @@ const MAX_SUPPORTED_JOINTS = 256
 export class TextureShader extends ShadowShader {
   private _jointMatrixTexture: StandardMaterialMatrixTexture
 
-  get maxSupportedJoints() {
-    return MAX_SUPPORTED_JOINTS
-  }
-
-  static isSupported(renderer: PIXI.Renderer) {
+  static isSupported(renderer: Renderer) {
     return StandardMaterialMatrixTexture.isSupported(renderer)
   }
 
-  constructor(renderer: PIXI.Renderer) {
+  constructor(renderer: Renderer) {
     super(renderer, [
       "USE_SKINNING 1", "USE_SKINNING_TEXTURE 1", "MAX_JOINT_COUNT " + MAX_SUPPORTED_JOINTS
     ])
@@ -30,11 +25,11 @@ export class TextureShader extends ShadowShader {
   createShaderGeometry(geometry: MeshGeometry3D) {
     let result = super.createShaderGeometry(geometry)
     if (geometry.joints) {
-      result.addAttribute("a_Joint1", new PIXI.Buffer(geometry.joints.buffer),
+      result.addAttribute("a_Joint1", new Buffer(geometry.joints.buffer),
         4, false, geometry.joints.componentType, geometry.joints.stride)
     }
     if (geometry.weights) {
-      result.addAttribute("a_Weight1", new PIXI.Buffer(geometry.weights.buffer),
+      result.addAttribute("a_Weight1", new Buffer(geometry.weights.buffer),
         4, false, geometry.weights.componentType, geometry.weights.stride)
     }
     return result